GWM/requirements.txt
altair==5.5.0
ansi2html==1.9.2
asgiref==3.8.1
attrs==25.3.0
autopep8==1.6.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
channels==4.2.2
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dash==2.11.1
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dj-database-url==0.5.0
Django==5.1.7
django-environ==0.8.1
django-grappelli==4.0.2
django-jazzmin==3.0.1
django-plotly-dash==2.4.6
django-widget-tweaks==1.5.0
dpd-static-support==0.0.5
dpd_components==0.1.0
Flask==2.2.5
gitdb==4.0.12
GitPython==3.1.44
gunicorn==20.1.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
narwhals==1.34.1
nest-asyncio==1.6.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyarrow==19.0.1
pycodestyle==2.8.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2021.3
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.24.0
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sqlparse==0.4.2
streamlit==1.44.1
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==2.2.3
whitenoise==5.3.0
zipp==3.21.0

#pip install dpd-static-support - Pode precisar caso não esteja aí, só pra garantir